Add TSV processing mode
When no file arguments are specified input is read from stdin and it is
presumed to be in TSV format.  It is also possible to output TSV format
using the option --tsv.  This enables filtering of tags using an
external program, e.g.

    tag --tsv *.flac | awk 'BEGIN{FS=OFS="\t"} { ... }' | tag

Note that this type of scripting will silently fail if tags or paths
contain tab characters.